Move initializers and restructure application-loader.tsx
