+++ src/components/application-loader/initializers/i18n.ts
